+++ lib/bump_ext/__init__.py
@@ -51,4 +51,4 @@ __all__ = [
     "PipelineDB",
 ]
 
-SCHEMA_VERSION = "0.0.4"
+SCHEMA_VERSION = "0.0.5"